# egld-connector/module/openapi.yaml
openapi: 3.0.0
servers:
  - url: 'https://api-eu1.tatum.io'
info:
  version: 1.0.0
  title: Tatum Elrond Network (EGLD) API
...
paths:
  /v3/egld/wallet:
    get:
      description: |
        <h4>1 credit per API call.</h4><br/>
        <p>The Elrond Address format is bech32, specified by the BIP 0173. The address always starts with an erd1. It is very convenient and secure, since it can generate 2^31 addresses from 1 mnemonic phrase.
        Mnemonic phrase consists of 24 special words in defined order and can restore access to all generated addresses and private keys.
        <br/>
        Each address is identified by 3 main values:
        <ul><li>Private Key - your secret value, which should never be revealed</li>
        <li>Public Key - public address to be published</li>
        <li>Derivation index - index of generated address</li></ul>
        </p>
        <p>Tatum follows BIP44 specification and generates for EGLD wallet with derivation path m'/44'/508'/0'/0'.
        More about BIP44 HD wallets can be found here - <a target="_blank" href="https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki">https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki</a>.
        Generate BIP44 compatible EGLD wallet.</p>
      operationId: EgldGenerateWallet
      parameters:
        - description: "Mnemonic to use for generation of private key."
          in: query
          name: mnemonic
          required: false
          schema:
            type: string
            maxLength: 500
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  mnemonic:
                    type: string
                    description: Mnemonic to use for generation of private key
                    example: "quantum tobacco key they maid mean crime youth chief jungle mind design broken tilt bus shoulder leaf good forward erupt split divert bread kitten"
          description: OK
        400: &API400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
          description: Bad Request. Validation failed for the given object in the HTTP Body or Request parameters.
        401: &API401
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error401NotActive"
                  - $ref: "#/components/schemas/Error401Invalid"
          description: Unauthorized. Not valid or inactive subscription key present in the HTTP Header.
        500: &API500
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500"
          description: Internal server error. There was an error on the server during the processing of the request.
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Generate EGLD wallet
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/address/{mnemonic}/{index}:
    get:
      description: |
        <h4>1 credit per API call.</h4><br/>
        <p>Generate EGLD account deposit address from mnemonic phrase. Deposit address is generated for the specific
        index - each mnemonic phrase can generate up to 2^31 addresses starting from index 0 until 2^31.</p>
      operationId: EgldGenerateAddress
      parameters:
        - description: "Mnemonic to use for generation of address."
          in: path
          name: mnemonic
          required: true
          schema:
            type: string
            example: "quantum tobacco key they maid mean crime youth chief jungle mind design broken tilt bus shoulder leaf good forward erupt split divert bread kitten"
        - description: "Derivation index of desired address to be generated."
          in: path
          name: index
          required: true
          schema:
            type: string
            example: 0
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                    description: EGLD address
                    example: erd17k95m339aqzxzyvjjjfa3lka0yyeqgcsda50tw5z9g73ycfe2caq9e6jq7
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Generate EGLD account address from mnemonic
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/wallet/priv:
    post:
      description: |
        <h4>1 credit per API call.</h4><br/>
        <p>Generate private key of address from mnemonic for given derivation path index. Private key is generated for the specific index - each mnemonic
        can generate up to 2^31 private keys starting from index 0 until 2^31.</p>
      operationId: EgldGenerateAddressPrivateKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivKeyRequest"
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivKey"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Generate EGLD private key
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/node/{xApiKey}/*:
    post:
      description: |
        <h4>2 credits per API call.</h4><br/>
        <p>Use this endpoint URL as a http-based driver to connect directly to the EGLD node provided by Tatum.
        To learn more about EGLD, please visit <a href="https://docs.elrond.com/sdk-and-tools/rest-api/nodes/ target="_blank">EGLD developer's guide.</a></p>
      example: "/v3/egld/node/{xApiKey}/transaction/cost" 
      operationId: EgldNodePost
      parameters:
        - description: "Tatum X-API-Key used for authorization."
          in: path
          name: xApiKey
          required: true
          schema:
            type: string
            example: asdlkfjnqunalkwjf124387ad
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Any valid method content body.
              example: {"version": 1,"chainId": "v1.0.141","nonce": 42,"value": "100000000000000000","receiver": "erd1cux02zersde0l7hhklzhywcxk4u9n4py5tdxyx7vrvhnza2r4gmq4vw35r","sender": "erd1njqj2zggfup4nl83x0nfgqjkjserm7mjyxdx5vzkm8k0gkh40ezqtfz9lg","gasPrice": 1000000000,"gasLimit": 70000,"data": "food for cats","signature": "93207c579bf57be03add632b0e1624a73576eeda8a1687e0fa286f03eb1a17ffb125ccdb008a264c402f074a360442c7a034e237679322f62268b614e926d10f"}
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                description: Any valid response.
                example: 77000
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      x-code-samples:
      summary: Node HTTP driver
      tags:
        - Blockchain / Elrond Network (EGLD)
    get:
      description: |
        <h4>2 credits per API call.</h4><br/>
        <p>Use this endpoint URL as a http-based driver to connect directly to the EGLD node provided by Tatum.
        To learn more about EGLD, please visit <a href="https://docs.elrond.com/sdk-and-tools/rest-api/nodes/ target="_blank">EGLD developer's guide.</a></p>
      example: "/v3/egld/node/{xApiKey}/network/config"
      operationId: EgldNodeGet
      parameters:
        - description: "Tatum X-API-Key used for authorization."
          in: path
          name: xApiKey
          required: true
          schema:
            type: string
            example: asdlkfjnqunalkwjf12341kljl
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                description: Any valid response.
                example: {"config":{"erd_chain_id":"1","erd_denomination":18,"erd_gas_per_data_byte":1500,"erd_latest_tag_software_version":"v1.1.0.0","erd_meta_consensus_group_size":400,"erd_min_gas_limit":50000,"erd_min_gas_price":1000000000,"erd_min_transaction_version":1,"erd_num_metachain_nodes":400,"erd_num_nodes_in_shard":400,"erd_num_shards_without_meta":3,"erd_round_duration":6000,"erd_shard_consensus_group_size":63,"erd_start_time":1596117600}}
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      x-code-samples:
      summary: Node HTTP driver
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/block/current:
    get:
      description: "<h4>1 credit per API call.</h4><br/>
        <p>Get EGLD current block number. This is the number of the latest block in the blockchain.</p>"
      operationId: EGldGetCurrentBlock
      responses:
        200:
          content:
            application/json:
              schema:
                type: number
                description: Current block number
                example: 649127
          description: OK
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get current block number
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/block/{hash}:
    get:
      description: "<h4>1 credit per API call.</h4><br/>
        <p>Get EGLD block by block hash or block number. <a href='https://docs.elrond.com/sdk-and-tools/rest-api/blocks/' target='_blank'> EGLD docs </a></p>"
      operationId: EgldGetBlock
      parameters:
        - description: "Block hash or nonce"
          in: path
          name: hash or nonce
          required: true
          schema:
            type: string
            example: a6f2ac15a6b1bafdde9afff2297cef49c4c523c516f8ee12fed54be070e9512b
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EgldBlock"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get EGLD block by hash
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/account/balance/{address}:
    get:
      description: "<h4>1 credit per API call.</h4><br/>
        <p>Get account balance in EGLD.</p>"
      operationId: EgldGetBalance
      parameters:
        - description: "Account address"
          in: path
          name: address
          required: true
          schema:
            type: string
            example: erd17k95m339aqzxzyvjjjfa3lka0yyeqgcsda50tw5z9g73ycfe2caq9e6jq7
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: string
                    description: Balance in EGLD
                    example: "10.52"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get EGLD Account balance
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/transaction/{hash}:
    get:
      description: "<h4>1 credit per API call.</h4><br/>
        <p>Get EGLD transaction by transaction hash. Detail result please find here <a href='https://docs.elrond.com/sdk-and-tools/rest-api/transactions/#get-transaction' target='_blank'> EGLD docs </a></p>"
      operationId: EgldGetTransaction
      parameters:
        - description: "Transaction hash"
          in: path
          name: hash
          required: true
          schema:
            type: string
            example: 99996224823736c1e9b8484ed74c1573049478f871d6f94b86811fb1c7b2addd
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EgldTx"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        403:
          description: Forbidden. The request is authenticated, but it is not possible to required perform operation due to logical error or invalid permissions.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error403EgldBroadcast"
                  - $ref: "#/components/schemas/Error403EgldGasCalculation"
                  - $ref: "#/components/schemas/Error403EgldTxHash"
                  - $ref: "#/components/schemas/Error403EgldTxSign"
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get EGLD Transaction
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/transaction/address/:address:
    get:
      description: |
        <h4>1 credit per API call.</h4><br/>
        <p>This endpoint allows one to retrieve the latest 20 transactions sent from an address.</p>
      operationId: EgldGetTransactionAddress
      parameters:
        - description: "address"
          in: path
          name: address
          required: true
          schema:
            type: string
            maxLength: 62
            minLength: 62
            example: erd17k95m339aqzxzyvjjjfa3lka0yyeqgcsda50tw5z9g73ycfe2caq9e6jq7
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                description: List of transactions sent from an address.
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get count of outgoing EGLD transactions
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/transaction/count/{address}:
    get:
      description: |
        <h4>1 credit per API call.</h4><br/>
        <p>Get a number of outgoing EGLD transactions for the address. When a transaction is sent, there can be multiple outgoing transactions,
        which are not yet processed by the blockchain. To distinguish between them, there is a counter called a nonce, which represents
        the order of the transaction in the list of outgoing transactions.</p>
      operationId: EgldGetTransactionCount
      parameters:
        - description: "address"
          in: path
          name: address
          required: true
          schema:
            type: string
            maxLength: 62
            minLength: 62
            example: erd17k95m339aqzxzyvjjjfa3lka0yyeqgcsda50tw5z9g73ycfe2caq9e6jq7
      responses:
        200:
          content:
            application/json:
              schema:
                type: number
                example: 5
                description: Number of transactions, that were made from this address. Can be used as a nonce parameter in send transaction calls.
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Get count of outgoing EGLD transactions
      tags:
        - Blockchain / Elrond Network (EGLD)
  /v3/egld/transaction:
    post:
      description: |
        <h4>2 credits per API call.</h4><br/>
        <p>Send EGLD from account to account.<br/><br/>
        This operation needs the private key of the blockchain address. Every time the funds are transferred, the transaction must be signed with the corresponding private key.
        No one should ever send it's own private keys to the internet because there is a strong possibility of stealing keys and loss of funds. In this method, it is possible to enter privateKey
        or signatureId. PrivateKey should be used only for quick development on testnet versions of blockchain when there is no risk of losing funds. In production,
        <a href="https://github.com/tatumio/tatum-kms" target="_blank">Tatum KMS</a> should be used for the highest security standards, and signatureId should be present in the request.
        Alternatively, using the Tatum client library for supported languages or Tatum Middleware with a custom key management system is possible.
        </p>
      operationId: EgldBlockchainTransfer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/TransferEgldBlockchain"
                - $ref: "#/components/schemas/TransferEgldBlockchainKMS"
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TransactionHashKMS"
                  - $ref: "#/components/schemas/SignatureId"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        403:
          description: Forbidden. The request is authenticated, but it is not possible to required perform operation due to logical error or invalid permissions.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error403EgldBroadcast"
                  - $ref: "#/components/schemas/Error403EgldGasCalculation"
                  - $ref: "#/components/schemas/Error403EgldTxHash"
                  - $ref: "#/components/schemas/Error403EgldTxSign"
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Send EGLD from account to account
      tags:
        - Blockchain / Elrond Network (EGLD)
...
  /v3/egld/broadcast:
    post:
      description: |
        <h4>2 credits per API call.</h4><br/>
        <p>Broadcast signed transaction to EGLD blockchain. This method is used internally from Tatum KMS or Tatum client libraries.
        It is possible to create custom signing mechanism and use this method only for broadcasting data to the blockchian.</p>
      operationId: EgldBroadcast
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastKMS"
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHashKMS"
          description: OK
        400:
          <<: *API400
        401:
          <<: *API401
        403:
          description: Forbidden. The request is authenticated, but it is not possible to required perform operation due to logical error or invalid permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error403EgldBroadcast"
        500:
          <<: *API500
      security:
        - X-Api-Key: [ ]
      x-code-samples:
      summary: Broadcast signed EGLD transaction
      tags:
        - Blockchain / Elrond Network (EGLD)
components:
  securitySchemes:
    X-Api-Key:
      description: |
        To maintain the highest security level, Tatum requires X-API-Key HTTP header.
        This X-API-Key represents identification of the plan and can be used by 1 developer or per environment.
        Each X-API-Key is bound to the test or production blockchains.
      in: header
      name: x-api-key
      type: apiKey
  schemas:
    EgldBlock:
      type: object
      properties:
        nonce:
          description: The block height.
          type: number
          example: 5440076
        round:
          description: The round number.
          type: number
          example: 186582
        hash:
          description: Hash of the block.
          type: string
          example: bd2de3618929b696807f6ef8a619f93d29d639aec9277f9c6f8569a8487141b1
        prevBlockHash:
          description: Hash of the previous block.
          type: string
          example: 2a11aad21e933be58200839ff875f0ca7e955b8c483a0e117cb92409e5d2a5da
        epoch:
          description: An epoch is a sequence of consecutive rounds during which the configuration of the network does not change (currently aprox. 24 hrs in length).
          type: number
          example: 18
        numTxs:
          description: Number of transactions in current block.
          type: number
          example: 8
        shardBlocks:
          description: Array of shard blocks
          items:
            $ref: "#/components/schemas/EgldShardBlock"
        transactions:
          description: Array of transactions.
          type: array
          items:
            $ref: "#/components/schemas/EgldTx"
    EgldShardBlock:
      type: object
      properties:
        hash:
          description: The block hash.
          type: string
          example: a6f2ac15a6b1bafdde9afff2297cef49c4c523c516f8ee12fed54be070e9512b
        nonce:
          description: The block height.
          type: number
          example: 5440070
        shard:
          description: 
          type: number
          example: 1
...
    Error403EgldBroadcast:
      properties:
        errorCode:
          description: egld.broadcast.failed
          example: egld.broadcast.failed
          type: string
        message:
          description: Unable to broadcast transaction due to ${error}.
          example: Unable to broadcast transaction due to ${error}.
          type: string
        statusCode:
          description: "403"
          example: 403
          type: number
      required:
        - errorCode
        - message
        - statusCode
      type: object
    Error403EgldTxHash:
      properties:
        errorCode:
          description: egld.transaction.hash
          example: egld.transaction.hash
          type: string
        message:
          description: Unable to calculate transaction hash. ${error}
          example: Unable to calculate transaction hash. ${error}
          type: string
        statusCode:
          description: "403"
          example: 403
          type: number
      required:
        - errorCode
        - message
        - statusCode
      type: object
    Error403EgldTxSign:
      properties:
        errorCode:
          description: egld.transaction.sign
          example: egld.transaction.sign
          type: string
        message:
          description: Unable to sign transaction. ${error}
          example: Unable to sign transaction. ${error}
          type: string
        statusCode:
          description: "403"
          example: 403
          type: number
      required:
        - errorCode
        - message
        - statusCode
      type: object
    Error403EgldGasCalculation:
      properties:
        errorCode:
          description: egld.transaction.gas
          example: egld.transaction.gas
          type: string
        message:
          description: Unable to calculate gas limit for transaction. ${error}
          example: Unable to calculate gas limit for transaction. ${error}
          type: string
        statusCode:
          description: "403"
          example: 403
          type: number
      required:
        - errorCode
        - message
        - statusCode
      type: object
...